requirements.txt
pandas
numpy
matplotlib
scikit-learn
xgboost
scikit-optimize
sagemaker
boto3
python-dotenv
